requirements.txt
secure-smtplib
pandas
openpyxl
selenium
requests
python-dotenv